test(wallet): DRY contract tests
