+++ packages/wallet/contract/test/test-singleWallet.js
@@ -5,115 +5,20 @@ import { test as anyTest } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 import '@agoric/vats/src/core/types.js';
 import '@agoric/zoe/exported.js';
 
-import { makeIssuerKit } from '@agoric/ertp';
-import { connectFaucet } from '@agoric/inter-protocol/src/proposals/demoIssuers.js';
 import { unsafeMakeBundleCache } from '@agoric/swingset-vat/tools/bundleTool.js';
-import {
-  installBootContracts,
-  makeAddressNameHubs,
-  makeBoard,
-} from '@agoric/vats/src/core/basic-behaviors.js';
-import { setupClientManager } from '@agoric/vats/src/core/chain-behaviors.js';
-import {
-  makeAgoricNamesAccess,
-  makePromiseSpace,
-} from '@agoric/vats/src/core/utils.js';
-import {
-  makeChainStorageRoot,
-  makeStorageNode,
-} from '@agoric/vats/src/lib-chainStorage.js';
+import { makeStorageNode } from '@agoric/vats/src/lib-chainStorage.js';
 import { makeNameHubKit } from '@agoric/vats/src/nameHub.js';
-import { buildRootObject as boardRoot } from '@agoric/vats/src/vat-board.js';
-import { buildRootObject as mintsRoot } from '@agoric/vats/src/vat-mints.js';
-import { makeZoeKit } from '@agoric/zoe';
-import { makeFakeVatAdmin } from '@agoric/zoe/tools/fakeVatAdmin.js';
-import { makeLoopback } from '@endo/captp';
 import { E, Far } from '@endo/far';
 import path from 'path';
-import { devices } from './devices.js';
-import { subscriptionKey } from './supports.js';
+import { makeTestSpace, subscriptionKey } from './supports.js';
 
 /** @type {import('ava').TestInterface<Awaited<ReturnType<makeTestContext>>>} */
 // @ts-expect-error cast
 const test = anyTest;
 
-const setUpZoeForTest = async () => {
-  const { makeFar } = makeLoopback('zoeTest');
-  const { zoeService, feeMintAccess: nonFarFeeMintAccess } = makeZoeKit(
-    makeFakeVatAdmin(() => {}).admin,
-  );
-  /** @type {import('@endo/far').ERef<ZoeService>} */
-  const zoe = makeFar(zoeService);
-  const feeMintAccess = await makeFar(nonFarFeeMintAccess);
-  return {
-    zoe,
-    feeMintAccess,
-  };
-};
-harden(setUpZoeForTest);
-
-export const mockChainStorageRoot = () => {
-  const toStorage = v => v;
-  return makeChainStorageRoot(toStorage, 'swingset', 'mockChainStorageRoot');
-};
-
 const makeTestContext = async t => {
-  const space = /** @type {any} */ (makePromiseSpace(t.log));
-  const { consume, produce } =
-    /** @type { BootstrapPowers & { consume: { loadVat: (n: 'mints') => MintsVat }} } */ (
-      space
-    );
-  const { agoricNames, spaces } = makeAgoricNamesAccess();
-  produce.agoricNames.resolve(agoricNames);
-
-  const { zoe, feeMintAccess } = await setUpZoeForTest();
-  produce.zoe.resolve(zoe);
-  produce.feeMintAccess.resolve(feeMintAccess);
-
-  produce.loadVat.resolve(name => {
-    switch (name) {
-      case 'mints':
-        return mintsRoot();
-      case 'board':
-        return boardRoot();
-      default:
-        throw Error('unknown loadVat name');
-    }
-  });
-
-  const bldKit = makeIssuerKit('BLD');
-  produce.bldIssuerKit.resolve(bldKit);
-  produce.chainStorage.resolve(mockChainStorageRoot());
-
-  produce.bankManager.resolve(
-    Promise.resolve(
-      Far('mockBankManager', {
-        // @ts-expect-error
-        getBankForAddress: _a =>
-          Far('mockBank', {
-            getPurse: () => ({
-              deposit: async (_, _x) => {
-                return null;
-              },
-            }),
-            getAssetSubscription: () => null,
-          }),
-      }),
-    ),
-  );
-
-  const vatPowers = {
-    D: x => x,
-  };
-
-  await Promise.all([
-    // @ts-expect-error
-    makeBoard({ consume, produce, ...spaces }),
-    makeAddressNameHubs({ consume, produce, ...spaces }),
-    installBootContracts({ vatPowers, devices, consume, produce, ...spaces }),
-    setupClientManager({ consume, produce, ...spaces }),
-    connectFaucet({ consume, produce, ...spaces }),
-  ]);
+  const { consume } = await makeTestSpace(t.log);
+  const { agoricNames, zoe } = consume;
 
   // Adapted from perAddress in makeAddressNameHubs()
   const reserveAddress = address => {
